Refactor screenCurtain to be a package with an internal _screenCurtain module containing all internal API
